Split data/instructions.yml into separate files.
I also added a bit of data, which is what inspired me to split the
content into separate files.  I think this should be a little easier
to work with.

Oh, and the output format changed a bit; instead of a list of families
with a "name" property for each one, the outermost object is now a map
and the name has moved to the key.  I don't think it's a big deal, but
it makes navigation a little easier.